edit entry (save) in frontend

+++ Component/branches/SobiPro-1.1/Site/opt/fields/image.php
@@ -2,21 +2,17 @@
 /**
  * @version: $Id$
  * @package: SobiPro Component for Joomla!
-
  * @author
  * Name: Sigrid Suski & Radek Suski, Sigsiu.NET GmbH
  * Email: sobi[at]sigsiu.net
  * Url: http://www.Sigsiu.NET
-
  * @copyright Copyright (C) 2006 - 2013 Sigsiu.NET GmbH (http://www.sigsiu.net). All rights reserved.
  * @license GNU/GPL Version 3
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License version 3
  * as published by the Free Software Foundation, and under the additional terms according section 7 of GPL v3.
  * See http://www.gnu.org/licenses/gpl.html and http://sobipro.sigsiu.net/licenses.
-
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
-
  * $Date$
  * $Revision$
  * $Author$
@@ -259,7 +255,7 @@ class SPField_Image extends SPField_Inbox implements SPFieldInterface
 		if ( !( $this->enabled ) ) {
 			return false;
 		}
-			$del = SPRequest::bool( $this->nid . '_delete', false, $request );
+		$del = SPRequest::bool( $this->nid . '_delete', false, $request );
 		static $store = null;
 		$cache = false;
 		if ( $store == null ) {
@@ -415,6 +411,9 @@ class SPField_Image extends SPField_Inbox implements SPFieldInterface
 		if ( is_array( $files ) && count( $files ) ) {
 			SPLoader::loadClass( 'cms.base.fs' );
 			foreach ( $files as $file ) {
+				if ( !( strlen( $file ) ) ) {
+					continue;
+				}
 				$file = Sobi::FixPath( SOBI_ROOT . "/{$file}" );
 				if ( SPFs::exists( $file ) ) {
 					SPFs::delete( $file );
